AG-8627 Move tests into a new zoomOnDataChange.test.ts file
